refactor: folder cleanup
