+++ packages/frontend/src/redux/reducerStatus/initialState/initialErrorState.js
