+++ packages/frontend/src/redux/slices/availableAccounts/index.js
@@ -3,7 +3,7 @@ import set from 'lodash.set';
 import { createSelector } from 'reselect';
 
 import handleAsyncThunkStatus from '../../reducerStatus/handleAsyncThunkStatus';
-import initialStatusState from '../../reducerStatus/initialStatusState';
+import initialStatusState from '../../reducerStatus/initialState/initialStatusState';
 import refreshAccountOwner from '../../sharedThunks/refreshAccountOwner';
 
 const SLICE_NAME = 'availableAccounts';
